Multi-ledger test: Merge files
